+++ examples/resources/aiven_byoc_permissions/resource.tf
@@ -1,0 +1,7 @@
+resource "aiven_byoc_permissions" "example" {
+  organization_id             = data.aiven_organization.main.id
+  custom_cloud_environment_id = aiven_byoc_aws_entity.example.custom_cloud_environment_id
+
+  accounts = [data.aiven_organization.main.id]
+  projects = ["my-project"]
+}