+++ CHANGELOG.md
@@ -27,6 +27,7 @@ nav_order: 1
 - Add `aiven_kafka_topic` field `config.message_timestamp_before_max_ms`: The maximum difference allowed between
   the timestamp when a broker receives a message and the timestamp specified in the message. Applies only for messages
   with timestamps earlier than the broker's timestamp.
+- Add `aiven_byoc_permissions` resource: Which accounts and projects can use a BYOC custom cloud environment to deploy services.
 
 ## [4.54.0] - 2026-04-09